Add files via upload
The Excel Dashboard is used to display overviews of large data tracks. Excel Dashboards use dashboard elements like tables, charts, and gauges to show the overviews. The dashboards ease the decision-making process by showing the vital parts of the data in the same window.
in this project, 
we have used Pivot Tables, Pivot Charts, and Slicers.
Fusion of Dashboard and Prepared Charts.
Added Slicers for Interactivity